+++ src/svf/downloader.ts
@@ -56,7 +56,7 @@ export class Downloader {
             const guidDir = path.join(urnDir, guid);
             fse.ensureDirSync(guidDir);
             const svf = await this.modelDerivativeClient.getDerivative(urn, derivative.urn);
-            fse.writeFileSync(path.join(guidDir, 'output.svf'), svf);
+            fse.writeFileSync(path.join(guidDir, 'output.svf'), new Uint8Array(svf));
             const reader = await SvfReader.FromDerivativeService(urn, guid, this.auth);
             const manifest = await reader.getManifest();
             for (const asset of manifest.assets) {